+++ docs/CHANGELOG.md
@@ -3,6 +3,9 @@ Version 3.28.1 (January 3, 2022)
 
 **User-visible changes:**
 
+Dropped support for `-ApermitUnsupportedJdkVersion` command-line argument.
+You can now run the Checker Framework under any JDK version, without a warning.
+
 **Implementation details:**
 
 **Closed issues:**